requirements.txt
<<<<<<< HEAD
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.19.0
pillow>=8.0.0
opencv-python>=4.5.0
tensorflow>=2.5.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
flask-sqlalchemy>=3.0.0
=======
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.19.0
pillow>=8.0.0
opencv-python>=4.5.0
tensorflow>=2.5.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
flask-sqlalchemy>=3.0.0
>>>>>>> 518cca968ba8bfdd0ffde9866e726a735025351d
flask-migrate>=4.0.0